requirements.txt
astroid==2.4.2
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.3.1
dateparser==1.0.0
Django==2.2
django-admin-interface==0.14.1
django-admin-notifications==0.6.3
django-colorfield==0.3.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jazzmin==2.4.4
django-model-utils==4.1.1
django-notifications-hq==1.6.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
python-binance==0.7.5
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
toml==0.10.2
txaio==20.12.1
tzlocal==2.1
urllib3==1.26.2
wrapt==1.12.1
zope.interface==5.2.0
